Move exception to common module
