+++ shardingsphere-control-panel/shardingsphere-orchestration/shardingsphere-orchestration-repository/shardingsphere-orchestration-repository-provider/shardingsphere-orchestration-repository-zookeeper-curator/src/test/java/org/apache/shardingsphere/orchestration/repository/zookeeper/handler/CuratorZookeeperExceptionHandlerTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.orchestration.repository.zookeeper.handler;
 
-import org.apache.shardingsphere.orchestration.repository.api.exception.OrchestrationException;
+import org.apache.shardingsphere.orchestration.repository.common.exception.OrchestrationException;
 import org.apache.zookeeper.KeeperException;
 import org.junit.Assert;
 import org.junit.Test;
